数据表与后端数据中价格转化为整数

+++ kernel/models/dto/tr_query_result.go
@@ -3,5 +3,5 @@ package dto
 type TrQueryResult struct {
 	Items        []*TransactionRecordDto `json:"items"`
 	Total        int64                   `json:"total"`
-	TrStatistics map[string]float64      `json:"trStatistics"`
+	TrStatistics map[string]int64        `json:"trStatistics"`
 }
